refactor: renamed folders and small tweaks
