+++ tests/tools/geoFirestoreCallbackRegistration.test.ts
@@ -121,7 +121,7 @@ describe('GeoFirestore GeoCallbackRegistration Tests:', () => {
         onKeyExitedRegistration.cancel();
         cl.x('p3');
 
-        return geoFirestore.set('loc3', [-80, -80]);
+        return wait(100).then(() => geoFirestore.set('loc3', [-80, -80]));
       }).then(() => {
         cl.x('p4');
